Add meal list action
- Create MealController.php
- Create endpoint to get meals
- Create MealResource.php
- Fix cast for meal price
- Fix cast for order amount

+++ app/Models/Meal.php
@@ -17,6 +17,6 @@ class Meal extends Model
     ];
 
     protected $casts = [
-        'price' => 'decimal',
+        'price' => 'float',
     ];
 }
